requirements.txt
python-dotenv~=1.0.0
requests~=2.31.0
discord.py~=2.3.2
aiofiles~=23.2.1
openai~=1.3.5
hydra-core~=1.3.2